[Backport 1.x] #27 #28 backport and bump to version 1.1.2 (#29)
* Added precision for codecov (#17)

* Added precision for codecov


* Add standard release pipeline library (#11)

* Add standard release pipeline library


* Create publishToNpm library (#21)

* Add npm publishing lib


* Add standard release pipeline library with generic trigger (#22)



* Add release workflow and readme (#23)

* Add release.yml


* Fix releasing.md (#25)



* Fix credential type for github bot (#26)

* Fix credential type for github bot


* Add untriaged label to new github issues (#27)



* Remove docker check for windows gradle check (#28)

* Remove docker check for windows gradle check


* add test results



* Remove docker check for windows gradle check (#28)

* Remove docker check for windows gradle check


* add test results



* upgrade to 1.1.1 with changes in #28


* Test results

+++ vars/runGradleCheck.groovy
@@ -7,7 +7,7 @@
  * compatible open source license.
  */
 void call(Map args = [:]) {
-    def lib = library(identifier: 'jenkins@1.1.1', retriever: legacySCM(scm))
+    def lib = library(identifier: 'jenkins@1.1.2', retriever: legacySCM(scm))
     def git_repo_url = args.gitRepoUrl ?: 'null'
     def git_reference = args.gitReference ?: 'null'
 
@@ -51,15 +51,17 @@ void call(Map args = [:]) {
                 ./gradlew --stop
                 rm -rf ~/.gradle
 
-                echo "Check existing dockercontainer"
-                docker ps -a
-                docker stop `docker ps -qa` > /dev/null 2>&1 || echo
-                docker rm `docker ps -qa` > /dev/null 2>&1 || echo
-                echo "Stop existing dockercontainer"
-                docker ps -a
+                if ! (uname -s | grep -i NT); then
+                    echo "Check existing dockercontainer"
+                    docker ps -a
+                    docker stop `docker ps -qa` > /dev/null 2>&1 || echo
+                    docker rm `docker ps -qa` > /dev/null 2>&1 || echo
+                    echo "Stop existing dockercontainer"
+                    docker ps -a
 
-                echo "Check docker-compose version"
-                docker-compose version
+                    echo "Check docker-compose version"
+                    docker-compose version
+                fi
 
                 echo "Check existing processes"
                 ps -ef | grep [o]pensearch | wc -l
